usb: Moved usb_get_state() into the main usb implementation

+++ src/platforms/common/cdcacm.h
@@ -28,12 +28,13 @@
 #ifndef __CDCACM_H
 #define __CDCACM_H
 
+#include <stdint.h>
 #include <stdbool.h>
 #include "usb.h"
 
 void usb_serial_set_config(usbd_device *dev, uint16_t wValue);
 /* Returns current usb configuration, or 0 if not configured. */
-int usb_get_config(void);
+uint16_t usb_get_config(void);
 
 bool gdb_uart_get_dtr(void);
